pyproject.toml
[project]
name = "lina"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "json5==0.12.0",
    "langchain==0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core==0.3.68",
    "langchain-milvus>=0.2.1",
    "langchain-openai==0.3.28",
    "langfuse>=3.2.0",
    "langgraph>=0.5.2",
    "langgraph-checkpoint>=2.1.0",
    "langgraph-checkpoint-sqlite>=2.0.10",
    "langgraph-supervisor==0.0.27",
    "networkx==3.5",
    "numpy==2.3.1",
    "openai==1.95.1",
    "pandas==2.3.1",
    "paramiko==3.5.1",
    "plotly==6.2.0",
    "psutil>=7.0.0",
    "pwinput==1.0.3",
    "pydantic==2.11.7",
    "pymilvus==2.5.10",
    "python-dotenv==1.1.1",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "scikit-learn==1.7.0",
    "scipy==1.16.0",
    "tenacity==9.1.2",
    "tqdm==4.67.1",
    "typer>=0.16.0",
    "uvicorn>=0.35.0",
    "watchdog==6.0.0",
]
